requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
altgraph==0.17.4
    # via pyinstaller
build==1.2.2.post1
    # via pip-tools
certifi==2024.8.30
    # via requests
charset-normalizer==3.4.0
    # via requests
click==8.1.7
    # via pip-tools
colorama==0.4.6
    # via
    #   build
    #   click
et-xmlfile==2.0.0
    # via openpyxl
greenlet==3.1.1
    # via playwright
idna==3.10
    # via requests
numpy==2.2.0
    # via pandas
openpyxl==3.1.5
    # via -r requirements.in
packaging==24.2
    # via
    #   build
    #   pyinstaller
    #   pyinstaller-hooks-contrib
pandas==2.2.3
    # via -r requirements.in
pefile==2023.2.7
    # via pyinstaller
pip-tools==7.4.1
    # via -r requirements.in
playwright==1.49.0
    # via -r requirements.in
pyee==12.0.0
    # via playwright
pyinstaller==6.11.1
    # via -r requirements.in
pyinstaller-hooks-contrib==2024.10
    # via pyinstaller
pymssql==2.3.2
    # via -r requirements.in
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.0.1
    # via -r requirements.in
pytz==2024.2
    # via pandas
pywin32-ctypes==0.2.3
    # via pyinstaller
requests==2.32.3
    # via -r requirements.in
six==1.17.0
    # via python-dateutil
typing-extensions==4.12.2
    # via pyee
tzdata==2024.2
    # via pandas
urllib3==2.2.3
    # via requests
wheel==0.45.1
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
